# openapi/basement-ui.yaml
openapi: 3.1.0
info:
  title: basement API
  version: 0.1.0
  description: 'Public OSS contract for basement''s API surface — the interface
      between

      the React frontend and Go backend. Frontend types are generated from this

      spec via openapi-typescript; backend handlers implement against it.


      **Versioning:** This is a 0.x spec (allowed to break). Contract freezes at

      version 1.0.0 when basement reaches feature-complete admin scope.

      '
  license:
      name: MIT
  contact:
      email: noreply@example.com
servers:
- url: http://localhost:8080/api/v1
  description: Local development server
...
paths:
...
  /auth/elevate:
      post:
        operationId: elevateSession
        summary: Sudo-style mode elevation (ADR-0003)
        description: |
          Re-authenticate to bump the current session's mode from USER to
          ADMIN (15min TTL) or from ADMIN to ELEVATED (5min TTL). On
          success, mints a fresh __Host-basement_session cookie with the
          new mode + mode_expires_at claims and returns the granted
          mode + expiry in the body so the frontend can drive its
          countdown chip without parsing the cookie.

          USER → ELEVATED directly is rejected (400 INVALID_TARGET_MODE);
          callers must elevate to ADMIN first. OIDC-only users get
          403 OIDC_USER_ELEVATION_NOT_SUPPORTED — v1.2.0c wires the
          OIDC `prompt=login` challenge for them.
        tags:
        - Auth
        x-basement-since: '1.2.0a'
        security:
        - cookieAuth: []
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                - password
                - target_mode
                properties:
                  password:
                    type: string
                  target_mode:
                    type: string
                    enum: [admin, elevated]
        responses:
          '200':
            description: |
              Elevation successful (mode/mode_expires_at/mode_ttl_seconds set
              + cookie re-issued) OR the dispatcher pivoted an OIDC-only user
              into the OIDC flow (requires_oidc=true + start_url; no cookie
              issued yet — FE POSTs start_url to begin the OIDC challenge).
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    mode:
                      type: string
                      enum: [admin, elevated]
                    mode_expires_at:
                      type: integer
                      format: int64
                      description: Unix seconds when the mode reverts.
                    mode_ttl_seconds:
                      type: integer
                      format: int64
                    requires_oidc:
                      type: boolean
                      description: |
                        v1.2.0c: true when the dispatcher detected an
                        OIDC-only user. The FE should POST to start_url to
                        begin the OIDC step-up challenge instead of
                        treating this as an error.
                    start_url:
                      type: string
                      description: |
                        v1.2.0c: relative path the FE POSTs to begin the
                        OIDC step-up. Always /api/v1/auth/elevate/oidc/start.
          '400':
            description: INVALID_TARGET_MODE / INVALID_JSON
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '401':
            description: INVALID_PASSWORD / UNAUTHORIZED
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '403':
            description: OIDC_USER_ELEVATION_NOT_SUPPORTED
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '503':
            description: OIDC_NOT_CONFIGURED (OIDC-only user but server has no OIDC provider)
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
...
  /share/{token}/auth:
      post:
        operationId: shareAuth
        summary: Authenticate for password-protected share (v1.3+)
        description: Validates password for a password-protected share and sets a scoped session cookie. On success, the client can access /share/{token}/list and /share/{token}/get endpoints. Returns 400 if no password is required.
        tags:
        - Shares
        x-basement-since: '1.3'
        security: []
        parameters:
        - name: token
          in: path
          required: true
          description: Share token
          schema:
            type: string
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  password:
                    type: string
                    format: password
        responses:
          '200':
            description: Authenticated successfully, cookie set
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Authenticated
          '400':
            description: No password required for this share
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '401':
            description: Invalid password
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '404':
            description: Share not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '410':
            description: Share has been revoked or expired
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
      cookieAuth:
        type: apiKey
        in: cookie
        name: __Host-basement_session
        description: JWT session cookie issued after authentication. HttpOnly, Secure,
          SameSite=Strict.
  schemas:
      ErrorResponse:
        type: object
        required:
        - error
        properties:
          error:
            type: object
            required:
            - code
            - message
            properties:
              code:
                type: string
                example: DRIVER_UNSUPPORTED
              message:
                type: string
              details:
                type: object
                additionalProperties: true
...
security:
- cookieAuth: []